Renamed files/folders to match new project names
