+++ CorgiShop/wwwroot/index.html
@@ -4,12 +4,12 @@
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
-    <title>CorgiShop.App</title>
+    <title>CorgiShop</title>
     <base href="/" />
     <link href="css/bootstrap/bootstrap.min.css" rel="stylesheet" />
     <link href="css/app.css" rel="stylesheet" />
     <link rel="icon" type="image/png" href="favicon.png" />
-    <link href="CorgiShop.App.styles.css" rel="stylesheet" />
+    <link href="CorgiShop.styles.css" rel="stylesheet" />
 </head>
 
 <body>
