+++ clean.sh
@@ -1,0 +1,1 @@
+find . -iname "bin" -o -iname "obj" | xargs rm -rf
